webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-ice.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-ice.html
-->

<meta charset=utf-8>
<title>WebRTC Test: the RTCPeerConnection Ice functionality</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var pc = null;
var iceCandidate = null;

var t = async_test(document.title);

function onIceChange2() {
  if (pc.iceConnectionState === "closed") {
    t.done();
  }
}

function addIceCandidateSuccess() {
  pc.oniceconnectionstatechange = onIceChange2;
  pc.close();
}

function addIceCandidateFailure() {
  t.step(function () {
    assert_unreached("addIceCandidateFailue was called.");
  });
  t.done();
}

function onIceChange1() {
  alert(11111);
  if (pc.iceConnectionState === "completed") {
    iceCandidate = new RTCIceCandidate({candidate:"nano nano"});

    t.step(function () {
      assert_throws(" SyntaxError", function () {
        pc.addIceCandidate(null, null, null);
      }, "addIceCandidate threw exception");
      assert_throws(" SyntaxError", function () {
        pc.addIceCandidate(iceCandidate, null, null);
      }, "addIceCandidate threw exception");
      assert_throws(" SyntaxError", function () {
        pc.addIceCandidate(iceCandidate, addIceCandidateSuccess, null);
      }, "addIceCandidate threw exception");
      pc.addIceCandidate(iceCandidate, addIceCandidateSuccess, addIceCandidateFailure);
    });
  }
}

test(function() {
  pc = new webkitRTCPeerConnection(null, null);
}, "Test new RTCPeerConnection do not throw exception");
pc.oniceconnectionstatechange = onIceChange1;
t.done();

</script>
